Fix #146349: 3D cursor still draws even though location is offscreen
Can happen if you set cursor location by snapping or manually.

In order to really skip drawing in `draw_output`, we have to set
`enabled_` to false in `begin_sync`.
This is already done for the "simple" cases covered by
`is_cursor_visible_XXX` but not for the "special" case where the cursor
is clipped in the view.

This is what this PR does.

Pull Request: https://projects.blender.org/blender/blender/pulls/146354