UI: Flip RNA icon + remove RNA_ADD
The RNA icon was rotating on the wrong direction.

Technically this is mostly a DNA helix, but even RNA can be a
double-helix sometimes. And both cases (RNA and DNA) are supposed to
have a right-hand orientation.

This PR also removes the RNA_ADD icon which is not used anywhere.

Based on suggestion by Brady Johnston.

Ref: !146600