Geometry Nodes: optimize Sample Grid Index node
This implements two optimizations for the Sample Grid Index node which caused a
lot of overhead. * Use an unsafe accessor which does not register itself with
the tree. This is necessary because we create many of these accessors in
parallel and they cause lots of thread contention. Using the unsafe accessor is
fine because we know that the grid won't change while we iterate over it. *
Access the grid when constructing the multi-function instead of in the `call`
function. This reduces similar thread contention when registering the tree
access token.

Overall, this results in a 4x speedup in my test with sampling neighboring
voxels.

The same optimization can probably be implemented for the Sample Grid node, but
that's not included here.

Pull Request: https://projects.blender.org/blender/blender/pulls/147271