Fix: newly created BMesh from Python left select_mode set to zero
The BMesh::selectmode is expected to be set when performing any
selection operations on the BMesh. While it wasn't possible to set
the select_mode to zero from Python, this meant assigning the attribute
from one BMesh to another could fail with an exception.
Aside from this, selection logic wouldn't work as expected.

Ref !147278